backend/app/requirements.txt
fastapi
uvicorn
pyjwt
asyncpg
